mv treebuilder prroc to processes.py
